ConsoleService refactoring.
