+++ hw01-xml-config/src/main/java/ru/otus/homework/common/ConsoleService.java
@@ -1,7 +1,15 @@
 package ru.otus.homework.common;
 
-public interface ConsoleService {
+import java.io.PrintStream;
+import lombok.RequiredArgsConstructor;
 
-  void println(String x);
+@RequiredArgsConstructor
+public class ConsoleService implements IOService {
 
+  private final PrintStream printStream;
+
+  @Override
+  public void println(String x) {
+    printStream.println(x);
+  }
 }
