+++ hw01-xml-config/src/main/java/ru/otus/homework/common/IOService.java
@@ -1,0 +1,7 @@
+package ru.otus.homework.common;
+
+public interface IOService {
+
+  void println(String x);
+
+}
