+++ hw01-xml-config/src/main/resources/spring-context.xml
@@ -1,18 +1,25 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <beans xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
   xmlns="http://www.springframework.org/schema/beans"
+  xmlns:util="http://www.springframework.org/schema/util"
   xsi:schemaLocation="http://www.springframework.org/schema/beans
-        http://www.springframework.org/schema/beans/spring-beans.xsd">
+        http://www.springframework.org/schema/beans/spring-beans.xsd
+        http://www.springframework.org/schema/util
+        http://www.springframework.org/schema/util/spring-util.xsd">
 
-  <bean class="ru.otus.homework.common.ConsoleServiceImpl" id="consoleService"/>
+  <bean class="ru.otus.homework.common.ConsoleService" id="IOService">
+    <constructor-arg name="printStream" ref="printStream"/>
+  </bean>
+
+  <bean class="ru.otus.homework.quiz.service.QuizServiceImpl" id="personService">
+    <constructor-arg name="quizDao" ref="quizDao"/>
+    <constructor-arg name="IOService" ref="IOService"/>
+  </bean>
 
   <bean class="ru.otus.homework.quiz.dao.QuizDaoCsv" id="quizDao">
     <constructor-arg name="defaultQuizCsv" value="quiz.csv"/>
   </bean>
 
-  <bean class="ru.otus.homework.quiz.service.QuizServiceImpl" id="personService">
-    <constructor-arg name="quizDao" ref="quizDao"/>
-    <constructor-arg name="consoleService" ref="consoleService"/>
-  </bean>
+  <util:constant id="printStream" static-field="java.lang.System.out"/>
 
 </beans>
